requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
insightface==0.7.3
onnxruntime==1.16.1
requests==2.31.0
supabase==2.18.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
threading-timer==1.0.0
pyjwt>=2.10.1